requirements.txt
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-PyMongo==2.3.0
marshmallow==3.22.0
python-dotenv==1.0.1
flasgger==0.9.7.1
pytest==8.3.3
pytest-flask==1.3.0
mongomock==4.2.0.post1
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.4